requirements.txt
beautifulsoup4==4.11.1
feedparser==6.0.8
greenlet==1.1.2
psycopg2-binary==2.9.3
pyaes==1.6.1
Pyrogram==2.0.17
PySocks==1.7.1
python-dotenv==0.20.0
sgmllib3k==1.0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
TgCrypto==1.2.3
